feat(hash-param): always update selected tab with hashParam
Previously the hashParam was updated only once; when the cosmoz-tabs element became visible. This is no longer compatible with the recent addition of `renderItem` support to cosmoz-data-nav, which does not destroy and re-render dataNav templates. Because views are re-used, the cross-queue selected tab synchronization does not work anymore.

I have tested and seen no errors caused by this change.